# preferences/src/main/openapi/api.yaml
openapi: 3.0.0
info:
  title: PSI Powierzenia - preferencje
  description: Detailed description can be found in the repository.
  version: 0.1.0
servers:
  - url: https://localhost/
    description: Main (production) server
  - url: http://localhost:3000
    description: Local server for development
paths:
  /preferences/{faculty}/{study}/{year}:
    get:
      summary: List preferences by lecturerId.
      description: Optional extended description in CommonMark or HTML.
      operationId: listAllPreferences
      parameters:
        - in: path
          name: faculty
          required: true
          schema:
            type: string
        - in: path
          name: study
          required: true
          schema:
            type: string
        - in: path
          name: year
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferenceList'
  /preferences:
    get:
      summary: List preferences by lecturerId.
      description: Optional extended description in CommonMark or HTML.
      operationId: listPreferences
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseList'
    post:
      summary: Add preferences.
      description: Optional extended description in CommonMark or HTML.
      operationId: addPreferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseList'
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseList'
    put:
      summary: Replace preferences.
      description: Optional extended description in CommonMark or HTML.
      operationId: replacePreferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseList'
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseList'
    delete:
      summary: Replace preferences.
      description: Optional extended description in CommonMark or HTML.
      operationId: deletePreferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseList'
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseList'
components:
  schemas:
    PreferenceList:
      type: array
      items:
        $ref: '#/components/schemas/Preference'
    Preference:
      type: object
      properties:
        lecturerId:
          $ref: '#/components/schemas/LecturerId'
        courseId:
          $ref: '#/components/schemas/CourseId'
    CourseList:
      type: array
      items:
        $ref: '#/components/schemas/CourseId'
    LecturerId:
      type: string
    CourseId:
      type: string